Move record tests to tests dir
